+++ src/main/java/com/troblecodings/signals/handler/ClientNameHandler.java
@@ -1,34 +1,26 @@
 package com.troblecodings.signals.handler;
 
 import java.nio.ByteBuffer;
-import java.util.Calendar;
 import java.util.HashMap;
 import java.util.Map;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 
 import com.troblecodings.core.interfaces.INetworkSync;
 import com.troblecodings.signals.core.ReadBuffer;
-import com.troblecodings.signals.tileentitys.BasicBlockEntity;
 
 import net.minecraft.client.Minecraft;
+import net.minecraft.client.multiplayer.ClientLevel;
 import net.minecraft.core.BlockPos;
-import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.entity.BlockEntity;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.network.NetworkEvent.ServerCustomPayloadEvent;
 
 public class ClientNameHandler implements INetworkSync {
 
     private static final Map<NameStateInfo, String> CLIENT_NAMES = new HashMap<>();
-    private static final ExecutorService SERVICE = Executors.newFixedThreadPool(5);
 
     public static String getClientName(final NameStateInfo info) {
         synchronized (CLIENT_NAMES) {
-            final String name = CLIENT_NAMES.get(info);
-            if (name == null)
-                return "";
-            return name;
+            return CLIENT_NAMES.getOrDefault(info, "");
         }
     }
 
@@ -46,24 +38,17 @@ public class ClientNameHandler implements INetworkSync {
         for (int i = 0; i < byteLength; i++) {
             array[i] = buffer.getByte();
         }
-        final Level world = mc.level;
         final String name = new String(array);
-        final long startTime = Calendar.getInstance().getTimeInMillis();
         synchronized (CLIENT_NAMES) {
             CLIENT_NAMES.put(new NameStateInfo(mc.level, pos), name);
         }
-        SERVICE.execute(() -> {
-            BlockEntity tile;
-            while ((tile = world.getBlockEntity(pos)) == null) {
-                final long currentTime = Calendar.getInstance().getTimeInMillis();
-                if (currentTime - startTime >= 10000) {
-                    return;
-                }
-                continue;
-            }
-            if (tile instanceof BasicBlockEntity) {
-                ((BasicBlockEntity) tile).setCustomName(name);
-            }
+        final ClientLevel world = mc.level;
+        mc.doRunTask(() -> {
+            final BlockState state = world.getBlockState(pos);
+            if (state == null)
+                return;
+            world.setBlocksDirty(pos, state, state);
+            world.setBlockAndUpdate(pos, state);
         });
     }