+++ src/main/java/com/troblecodings/signals/tileentitys/BasicBlockEntity.java
@@ -13,14 +13,12 @@ import com.troblecodings.signals.handler.NameStateInfo;
 import net.minecraft.core.BlockPos;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.world.level.block.entity.BlockEntity;
-import net.minecraft.world.level.block.state.BlockState;
 
 public class BasicBlockEntity extends BlockEntity implements NamableWrapper {
 
     public static final String GUI_TAG = "guiTag";
     public static final String POS_TAG = "posTag";
     protected final ArrayList<BlockPos> linkedPositions = new ArrayList<>();
-    protected String customName = null;
 
     public BasicBlockEntity(final TileEntityInfo info) {
         super(info.type, info.pos, info.state);
@@ -51,22 +49,13 @@ public class BasicBlockEntity extends BlockEntity implements NamableWrapper {
     @Override
     public String getNameWrapper() {
         final NameStateInfo info = new NameStateInfo(level, worldPosition);
-        if (customName == null || customName.isEmpty())
-            customName = level.isClientSide ? ClientNameHandler.getClientName(info)
-                    : NameHandler.getName(info);
-        return customName == null ? "" : customName;
+        return level.isClientSide ? ClientNameHandler.getClientName(info)
+                : NameHandler.getName(info);
+
     }
 
     @Override
     public boolean hasCustomName() {
-        if (customName == null)
-            getNameWrapper();
-        return customName != null;
-    }
-
-    public void setCustomName(final String name) {
-        this.customName = name;
-        final BlockState state = this.getBlockState();
-        this.level.setBlocksDirty(worldPosition, state, state);
+        return !getNameWrapper().isEmpty();
     }
 }